+++ qa/software/API_Test_Suite/src/caarray/client/test/full/FullApiFacade.java
@@ -376,7 +376,7 @@ public class FullApiFacade implements ApiFacade
     /* (non-Javadoc)
      * @see caarray.client.test.ApiFacade#copyMageTabZipToOutputStream(java.lang.String, gov.nih.nci.caarray.external.v1_0.CaArrayEntityReference, boolean)
      */
-    public Integer copyMageTabZipToOutputStream(String api,
+    public Long copyMageTabZipToOutputStream(String api,
             CaArrayEntityReference experimentReference, boolean compressed)
             throws Exception
     {
@@ -395,7 +395,7 @@ public class FullApiFacade implements ApiFacade
     /* (non-Javadoc)
      * @see caarray.client.test.ApiFacade#copyMageTabZipApiUtils(java.lang.String, gov.nih.nci.caarray.external.v1_0.CaArrayEntityReference, boolean)
      */
-    public Integer copyMageTabZipApiUtils(String api,
+    public Long copyMageTabZipApiUtils(String api,
             CaArrayEntityReference experimentReference, boolean compressed)
             throws Exception
     {
